// src/types/mock-data.type.ts
import { Coordinate } from './coordinate.type';

export type MockData = {
  titles: string[];
  descriptions: string[];
  cities: string[];
  previewImages: string[];
  offerImages: string[];
  names: string[];
  emails: string[];
  avatarPaths: string[];
  passwords: string[];
  coordinates: Coordinate[];
};